ingexai/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pytest
python-multipart
passlib[bcrypt]
python-jose
PyPDF2
python-docx
# ...add more as needed...
